♻️ ::  클래스명 UseCase 추가
