+++ yapaghetti-application/src/test/java/kr/hs/entrydsm/yapaghetti/domain/document/usecase/ApproveStayDocumentUseCaseTest.java
@@ -2,6 +2,7 @@ package kr.hs.entrydsm.yapaghetti.domain.document.usecase;
 
 import kr.hs.entrydsm.yapaghetti.domain.document.domain.Document;
 import kr.hs.entrydsm.yapaghetti.domain.document.domain.DocumentType;
+import kr.hs.entrydsm.yapaghetti.domain.document.spi.CommandDocumentPort;
 import kr.hs.entrydsm.yapaghetti.domain.document.spi.QueryDocumentPort;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,25 +15,33 @@ import java.util.UUID;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-class QueryPublicDocumentTest {
+class ApproveStayDocumentUseCaseTest {
 
     @Mock
     QueryDocumentPort queryDocumentPort;
 
+    @Mock
+    CommandDocumentPort commandDocumentPort;
+
     @InjectMocks
-    QueryPublicDocumentUseCase queryPublicDocumentUseCase;
+    ApproveStayDocumentUseCase approveStayDocumentUseCase;
 
     @Test
-    void queryPublicDocument() {
+    void approveDocument() {
         UUID documentId = UUID.randomUUID();
-        String content = "testContent";
+        UUID userId = UUID.randomUUID();
 
-        given(queryDocumentPort.queryDocumentByIdAndType(documentId, DocumentType.PUBLIC)).willReturn(
+        given(queryDocumentPort.queryDocumentByIdAndType(documentId, DocumentType.STAY)).willReturn(
                 Document.builder()
-                        .content(content)
+                        .id(documentId)
+                        .previewImagePath("testPreviewImagePath")
+                        .content("testContent")
+                        .type(DocumentType.STAY)
+                        .userId(userId)
+                        .isRejected(false)
                         .build()
         );
 
-        queryPublicDocumentUseCase.execute(documentId);
+        approveStayDocumentUseCase.execute(documentId);
     }
 }
