+++ yapaghetti-application/src/test/java/kr/hs/entrydsm/yapaghetti/domain/document/usecase/CopyPublicDocumentUseCaseTest.java
@@ -3,6 +3,7 @@ package kr.hs.entrydsm.yapaghetti.domain.document.usecase;
 import kr.hs.entrydsm.yapaghetti.domain.document.domain.Document;
 import kr.hs.entrydsm.yapaghetti.domain.document.domain.DocumentType;
 import kr.hs.entrydsm.yapaghetti.domain.document.spi.CommandDocumentPort;
+import kr.hs.entrydsm.yapaghetti.domain.document.spi.DocumentSecurityPort;
 import kr.hs.entrydsm.yapaghetti.domain.document.spi.QueryDocumentPort;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -15,7 +16,7 @@ import java.util.UUID;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-class RejectStayDocumentTest {
+class CopyPublicDocumentUseCaseTest {
 
     @Mock
     QueryDocumentPort queryDocumentPort;
@@ -23,18 +24,28 @@ class RejectStayDocumentTest {
     @Mock
     CommandDocumentPort commandDocumentPort;
 
+    @Mock
+    DocumentSecurityPort documentSecurityPort;
+
     @InjectMocks
-    RejectStayDocumentUseCase rejectStayDocumentUseCase;
+    CopyPublicDocumentUseCase copyPublicDocumentUseCase;
 
     @Test
-    void rejectDocument() {
-        UUID documentId = UUID.randomUUID();
+    void copy() {
+        UUID userId = UUID.randomUUID();
 
-        given(queryDocumentPort.queryDocumentByIdAndType(documentId, DocumentType.STAY)).willReturn(
+        given(documentSecurityPort.getCurrentUserId()).willReturn(userId);
+        given(queryDocumentPort.queryDocumentByUserIdAndType(userId, DocumentType.PUBLIC)).willReturn(
                 Document.builder()
+                        .previewImagePath("testPreviewImage")
+                        .content("testContent")
+                        .type(DocumentType.PUBLIC)
+                        .userId(userId)
+                        .isRejected(false)
                         .build()
         );
 
-        rejectStayDocumentUseCase.execute(documentId);
+        copyPublicDocumentUseCase.execute();
     }
+
 }
