+++ yapaghetti-application/src/test/java/kr/hs/entrydsm/yapaghetti/domain/document/usecase/CreateLocalDocumentUseCaseTest.java
@@ -17,7 +17,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-class CreateLocalDocumentTest {
+class CreateLocalDocumentUseCaseTest {
 
     @Mock
     CommandDocumentPort commandDocumentPort;
