+++ yapaghetti-application/src/test/java/kr/hs/entrydsm/yapaghetti/domain/document/usecase/QueryPublicDocumentUseCaseTest.java
@@ -2,7 +2,6 @@ package kr.hs.entrydsm.yapaghetti.domain.document.usecase;
 
 import kr.hs.entrydsm.yapaghetti.domain.document.domain.Document;
 import kr.hs.entrydsm.yapaghetti.domain.document.domain.DocumentType;
-import kr.hs.entrydsm.yapaghetti.domain.document.spi.DocumentSecurityPort;
 import kr.hs.entrydsm.yapaghetti.domain.document.spi.QueryDocumentPort;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -15,29 +14,25 @@ import java.util.UUID;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-public class QueryProtectedDocumentUrlTest {
+class QueryPublicDocumentUseCaseTest {
 
     @Mock
     QueryDocumentPort queryDocumentPort;
 
-    @Mock
-    DocumentSecurityPort documentSecurityPort;
-
     @InjectMocks
-    QueryProtectedDocumentUrlUseCase queryProtectedDocumentUrlUseCase;
+    QueryPublicDocumentUseCase queryPublicDocumentUseCase;
 
     @Test
-    void getUrl() {
-        UUID userId = UUID.randomUUID();
+    void queryPublicDocument() {
         UUID documentId = UUID.randomUUID();
+        String content = "testContent";
 
-        given(documentSecurityPort.getCurrentUserId()).willReturn(userId);
-        given(queryDocumentPort.queryDocumentByUserIdAndType(userId, DocumentType.PROTECTED)).willReturn(
+        given(queryDocumentPort.queryDocumentByIdAndType(documentId, DocumentType.PUBLIC)).willReturn(
                 Document.builder()
-                        .id(documentId)
+                        .content(content)
                         .build()
         );
 
-        queryProtectedDocumentUrlUseCase.execute();
+        queryPublicDocumentUseCase.execute(documentId);
     }
 }
