+++ yapaghetti-application/src/test/java/kr/hs/entrydsm/yapaghetti/domain/document/usecase/UpdateLocalDocumentUseCaseTest.java
@@ -19,7 +19,7 @@ import java.util.UUID;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-class UpdateLocalDocumentTest {
+class UpdateLocalDocumentUseCaseTest {
 
     @Mock
     DocumentQueryUserPort documentQueryUserPort;
