+++ yapaghetti-application/src/test/java/kr/hs/entrydsm/yapaghetti/domain/document/usecase/UpdateStayDocumentUseCaseTest.java
@@ -17,7 +17,7 @@ import java.util.UUID;
 import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
-class UpdateStayDocumentTest {
+class UpdateStayDocumentUseCaseTest {
 
     @Mock
     QueryDocumentPort queryDocumentPort;
